refactor: query-builder: Abstract duplicate code from Insert/Update/Delete
* Rename ModificationQueryBuilder to AbstractModifyQueryBuilder (UPDATE/DELETE)

* New AbstractPersistQueryBuilder for INSERT/UPDATE/DELETE
    * Common code for RETURNING expressions
    * Common code for INSERT/UPDATE value expressions

* Move UpdateDateColumn, VersionDateColumn, etc to value expression calculation
    * Matches InsertQueryBuilder behavior
